==> 02_scripts/02_fishbase_data.Rmd <==
---
title: "Data Collection and Manipulation"
author: "Lex Rosenberg"
date: "2023-06-30"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE, message=FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)

library(tidyverse)
library(here)
library(rfishbase)
library(naniar)
library(car)
library(sf)
library(sp)
library(ggpubr)
library(rstatix)
library(FSA)
```


##### Load data

```{r}
data <- read_csv(here::here("01_data", "01_raw_data", "homerange_pld.csv"))
```

##### Rename incorect species

```{r}
data <- data %>%
  mutate(species = str_replace(species, "Oblada melanuras", "Oblada melanurus")) %>%
  mutate(species = str_replace(species, "Pennahia aneas", "Pennahia aneus")) %>%
  mutate(species = str_replace(species, "Cromileptes altivelis", "Chromileptes altivelis"))
```

##### Make new clusters

```{r}
clusters <- read_csv(here::here("01_data", "02_processed_data", "clusters.csv")) %>%
  mutate(class2 = class) %>%
  mutate(class2 = case_when(
    class2 == 7 ~ 2,
    class2 == 1 ~ 3,
    TRUE ~ class2
  )) %>%
  mutate(class3 = case_when(
    class2 == 4 ~ 1,
    class2 == 2 ~ 2, 
    class2 == 8 ~ 3,
    class2 == 9 ~ 4,
    class2 == 6 ~ 5,
    class2 == 3 ~ 6
    ))

# rename incorrect species
clusters <- clusters %>%
  mutate(species = str_replace(species, "Oblada melanuras", "Oblada melanurus")) %>%
  mutate(species = str_replace(species, "Pennahia aneas", "Pennahia aneus")) %>%
  mutate(species = str_replace(species, "Cromileptes altivelis", "Chromileptes altivelis"))

# rename classes with descriptions: homerange/PLD
clusters <- clusters %>%
  mutate(class2 = case_when(
    class3 == 1 ~ "1: HR-mid/PLD-low",
    class3 == 2 ~ "2: HR-low/PLD-mid",
    class3 == 3 ~ "3: HR-mid/PLD-high",
    class3 == 4 ~ "4: HR-mid/PLD-veryhigh",
    class3 == 5 ~ "5: HR-mid/PLD-mid",
    class3 == 6 ~ "6: HR-high/PLD-mid"
  ))
```

##### Plot of new clusters

```{r}
data <- full_join(data, clusters)

fishbase_data_gmm <- data %>%
  mutate(class3 = as.factor(class3))

```

##### Load in taxonomic info

```{r}
fish <- load_taxa(server = "fishbase") %>% # load all taxonomic info from fishbase into fish
  as.data.frame() %>% 
  filter(Species %in% data$species) %>% 
  janitor::clean_names() %>% 
  distinct() %>% 
  select(species, spec_code, genus, order)

data <- left_join(data, fish)
```

##### Collect temperature data

```{r}
temp <- stocks() %>%
  filter(SpecCode %in% data$spec_code) %>%
  dplyr::select(SpecCode, TempMin, TempMax, TempPreferred) %>%
  mutate(temp_range = TempMax - TempMin) %>%
  arrange(SpecCode) %>% 
  distinct(SpecCode, .keep_all = TRUE)

temp <- drop_na(temp, TempPreferred)

temp <- janitor::clean_names(temp)

data <- left_join(data, temp)
```

##### Collect fecundity data

```{r}
fecundity <- fecundity() %>%
  filter(SpecCode %in% data$spec_code) %>%
  dplyr::select(SpecCode, FecundityMax, FecundityMin, FecundityMean) %>%
  group_by(SpecCode) %>%
  mutate(
    FecundityMax = mean(FecundityMax, na.rm = TRUE),
    FecundityMin = mean(FecundityMin, na.rm = TRUE),
    FecundityMean = mean(FecundityMean, na.rm = TRUE),
  ) %>%
  ungroup() %>%
  distinct() %>% 
  mutate_at(
    vars(FecundityMax, FecundityMin, FecundityMean), ~ ifelse(is.nan(.), NA, .)) %>% 
  mutate(
    FecundityMax = as.numeric(FecundityMax),
    FecundityMin = as.numeric(FecundityMin),
    FecundityMeanCalc = ((FecundityMax + FecundityMin)/2)
  ) %>%
  janitor::clean_names()


fecundity <- fecundity %>% # combine actual mean with calculated mean
  mutate(fecundity_mean = coalesce(fecundity_mean, fecundity_mean_calc)) %>%
  dplyr::select(-fecundity_mean_calc, -fecundity_min)


data <- left_join(data, fecundity)
```

##### Collect shape/age data

```{r}
species <- species() %>%
  filter(SpecCode %in% data$spec_code) %>%
  dplyr::select(SpecCode, BodyShapeI, LongevityWild) %>%
  janitor::clean_names() %>% 
  distinct()


data <- left_join(data, species)
```

##### Collect metabolic data

Only 74 species with oxygen consumption data:
```{r}
oxygen <- oxygen() %>%
  filter(SpecCode %in% data$spec_code) %>%
  dplyr::select(SpecCode, OxygenCons) %>%
  group_by(SpecCode) %>%
  summarise(OxygenCons = mean(OxygenCons)) %>%
  ungroup() %>%
  janitor::clean_names()

data <- left_join(data, oxygen)
```

##### Collect reproduction data
 
17 are non-planktonic of 298 species
```{r}
larvae <- larvae() %>%
  filter(SpecCode %in% data$spec_code) %>%
  dplyr::select(SpecCode, PlaceofDevelopment) %>%
  na.omit() %>%
  rename(dev_place = PlaceofDevelopment,
         spec_code = SpecCode) %>% 
  distinct() %>% 
  mutate(dev_place = case_when(
    dev_place != "planktonic" ~ "non-planktonic", 
    TRUE ~ dev_place))

data <- left_join(data, larvae)
```

##### Collect species richness by family data

```{r}
sp_richness <- load_taxa(server = "fishbase") %>%
  as.data.frame() %>% 
  filter(Family %in% data$family) %>%
  dplyr::select(Family, Species) %>% 
  distinct() %>%
  count(Family) %>%
  dplyr::rename(family = Family, sp_richness = n)

data <- left_join(data, sp_richness)
```

##### Fill IUCN codes

```{r}
iucn_codes <- read_csv(here::here("01_data", "01_raw_data", "iucn_codes.csv")) %>% 
  na.omit()

data <- data %>%
  left_join(iucn_codes, by = "species") %>%
  mutate(iucn_category = coalesce(iucn_category.x, iucn_category.y)) %>%
  dplyr::select(-iucn_category.x, -iucn_category.y)
```

##### Add common names

```{r}
names <- species(fields = c("SpecCode", "FBname")) %>%
  filter(SpecCode %in% data$spec_code) %>%
  rename(fishbase_name = FBname, spec_code = SpecCode) %>% 
  na.omit()

data <- left_join(data, names)
```


##### Look at missing data

```{r}
missing <- miss_var_summary(data)

missing_data <- data %>% 
  dplyr::select(
    oxygen_cons, temp_range, fecundity_mean, longevity_wild,
    temp_preferred, dev_place
  ) %>%
  rename(
    "Oxygen Consumption Rate" = oxygen_cons,
    "Temperature Range" = temp_range,
    "Average Fecundity" = fecundity_mean,
    "Maximum Recorded Lifespan" = longevity_wild,
    "Optimal Temperature" = temp_preferred,
    "Place of Development" = dev_place
  )

m1 <- gg_miss_var(missing_data)  +
  theme_bw(base_size = 16) +
  labs(y = "Original Missing Data", x = "") +
  theme(axis.title.x = element_text(vjust = -1))
```

##### Fill in missing data from genus averages

```{r}
# replace numeric NAs with mean of genus, categorical with mode of genus

mode <- function(x) {
  names(which.max(table(x)))
}

fish_genera <- load_taxa(server = "fishbase") %>%
  as.data.frame() %>% 
  filter(Genus %in% data$genus) %>%
  dplyr::select(SpecCode, Genus) %>%
  janitor::clean_names()


# Place of Development
missing_dev <- larvae() %>%
  filter(SpecCode %in% fish_genera$spec_code) %>%
  dplyr::select(SpecCode, PlaceofDevelopment) %>%
  rename(spec_code = SpecCode, 
         dev_place = PlaceofDevelopment) %>%
  na.omit() %>% 
  mutate(dev_place = case_when(
    dev_place != "planktonic" ~ "non-planktonic", 
    TRUE ~ dev_place))

missing_dev <- left_join(fish_genera, missing_dev, relationship = "many-to-many")

missing_dev <- missing_dev %>%
  mutate(dev_place = as.factor(dev_place)) %>%
  group_by(genus) %>%
  summarise(dev_place = mode(dev_place)) %>%
  ungroup()

genus_filled_data <- data %>%
  left_join(missing_dev, by = "genus") %>%
  mutate(dev_place = coalesce(dev_place.x, dev_place.y)) %>%
  dplyr::select(-dev_place.x, -dev_place.y)


# Longevity Wild
missing_long_1 <- species() %>%
  filter(SpecCode %in% fish_genera$spec_code) %>%
  dplyr::select(SpecCode, LongevityWild) %>%
  janitor::clean_names()

missing_long_1 <- left_join(fish_genera, missing_long_1, relationship = "many-to-many")

missing_long_1 <- missing_long_1 %>%
  mutate(longevity_wild = as.numeric(longevity_wild)) %>%
  na.omit() %>%
  group_by(genus) %>%
  summarise(longevity_wild = mean(longevity_wild)) %>%
  ungroup()

genus_filled_data <- genus_filled_data %>%
  left_join(missing_long_1, by = "genus") %>%
  mutate(longevity_wild = coalesce(longevity_wild.x, longevity_wild.y)) %>%
  dplyr::select(-longevity_wild.x, -longevity_wild.y)


# Preferred Temp
missing_temp_1 <- stocks() %>%
  filter(SpecCode %in% fish_genera$spec_code) %>%
  dplyr::select(SpecCode, TempPreferred) %>%
  janitor::clean_names()

missing_temp_1 <- left_join(fish_genera, missing_temp_1, relationship = "many-to-many")

missing_temp_1 <- missing_temp_1 %>%
  mutate(temp_preferred = as.numeric(temp_preferred)) %>%
  na.omit() %>%
  group_by(genus) %>%
  summarise(temp_preferred = mean(temp_preferred)) %>%
  ungroup()

genus_filled_data <- genus_filled_data %>%
  left_join(missing_temp_1, by = "genus") %>%
  mutate(temp_preferred.x = as.numeric(temp_preferred.x)) %>% 
  mutate(temp_preferred = coalesce(temp_preferred.x, temp_preferred.y)) %>%
  dplyr::select(-temp_preferred.x, -temp_preferred.y)


# Temp Range
missing_range_1 <- stocks() %>%
  filter(SpecCode %in% fish_genera$spec_code) %>%
  dplyr::select(SpecCode, TempMin, TempMax) %>%
  mutate(temp_range = TempMax - TempMin) %>%
  janitor::clean_names()

missing_range_1 <- left_join(fish_genera, missing_range_1, relationship = "many-to-many")

missing_range_1 <- missing_range_1 %>%
  mutate(temp_range = as.numeric(temp_range)) %>%
  na.omit() %>%
  group_by(genus) %>%
  summarise(temp_range = mean(temp_range)) %>%
  ungroup()

genus_filled_data <- genus_filled_data %>%
  left_join(missing_range_1, by = "genus") %>%
  mutate(temp_range = coalesce(temp_range.x, temp_range.y)) %>%
  dplyr::select(-temp_range.x, -temp_range.y)



# Fecundity Mean
missing_fec_1 <- fecundity() %>%
  filter(SpecCode %in% fish_genera$spec_code) %>%
  dplyr::select(SpecCode, FecundityMean) %>%
  janitor::clean_names()

missing_fec_1 <- left_join(fish_genera, missing_fec_1, relationship = "many-to-many")

missing_fec_1 <- missing_fec_1 %>%
  mutate(fecundity_mean = as.numeric(fecundity_mean)) %>%
  na.omit() %>%
  group_by(genus) %>%
  summarise(fecundity_mean = mean(fecundity_mean)) %>%
  ungroup()

genus_filled_data <- genus_filled_data %>%
  left_join(missing_fec_1, by = "genus") %>%
  mutate(fecundity_mean = coalesce(fecundity_mean.x, fecundity_mean.y)) %>%
  dplyr::select(-fecundity_mean.x, -fecundity_mean.y)


# Oxygen Consumption
missing_ox_1 <- oxygen() %>%
  filter(SpecCode %in% fish_genera$spec_code) %>%
  dplyr::select(SpecCode, OxygenCons) %>%
  janitor::clean_names()

missing_ox_1 <- left_join(fish_genera, missing_ox_1, relationship = "many-to-many")

missing_ox_1 <- missing_ox_1 %>%
  mutate(oxygen_cons = as.numeric(oxygen_cons)) %>%
  na.omit() %>%
  group_by(genus) %>%
  summarise(oxygen_cons = mean(oxygen_cons)) %>%
  ungroup()

genus_filled_data <- genus_filled_data %>%
  left_join(missing_ox_1, by = "genus") %>%
  mutate(oxygen_cons = coalesce(oxygen_cons.x, oxygen_cons.y)) %>%
  dplyr::select(-oxygen_cons.x, -oxygen_cons.y)

```


##### Asses missing data after genus fills - dev_place is complete

```{r}
genus_filled <- miss_var_summary(genus_filled_data)

# New missing plot
genus_filled_plot <- genus_filled_data %>% 
  dplyr::select(
    oxygen_cons, temp_range, fecundity_mean, longevity_wild,
    temp_preferred, dev_place
  ) %>%
  rename(
    "Oxygen Consumption Rate" = oxygen_cons,
    "Temperature Range" = temp_range,
    "Average Fecundity" = fecundity_mean,
    "Maximum Recorded Lifespan" = longevity_wild,
    "Optimal Temperature" = temp_preferred,
    "Place of Development" = dev_place
  )

m2 <- gg_miss_var(genus_filled_plot) + # figure showing the missing data
  theme_bw(base_size = 16) +
  labs(y = "Missing Data After Genus-Level Fill", x = "Variable") +
  theme(axis.title.x = element_text(vjust = -1))

```

##### Fill in missing data from family averages for remaining variables

```{r}
# replace numeric NAs with mean of family, categorical with mode of family

fish_fam <- load_taxa(server = "fishbase") %>%
  as.data.frame() %>% 
  filter(Family %in% data$family) %>%
  dplyr::select(SpecCode, Family) %>%
  janitor::clean_names()

# Longevity Wild
missing_long_2 <- species() %>%
  filter(SpecCode %in% fish_fam$spec_code) %>%
  dplyr::select(SpecCode, LongevityWild) %>%
  janitor::clean_names()

missing_long_2 <- left_join(fish_fam, missing_long_2)

missing_long_2 <- missing_long_2 %>%
  mutate(longevity_wild = as.numeric(longevity_wild)) %>%
  na.omit() %>%
  group_by(family) %>%
  summarise(longevity_wild = mean(longevity_wild)) %>%
  ungroup()

filled_data <- genus_filled_data

filled_data <- filled_data %>%
  left_join(missing_long_2, by = "family") %>%
  mutate(longevity_wild = coalesce(longevity_wild.x, longevity_wild.y)) %>%
  dplyr::select(-longevity_wild.x, -longevity_wild.y)


# Preferred Temp
missing_temp_2 <- stocks() %>%
  filter(SpecCode %in% fish_fam$spec_code) %>%
  dplyr::select(SpecCode, TempPreferred) %>%
  filter(SpecCode != 67902) %>%  # remove erroneous data (temperature of 105428 ºC)
  janitor::clean_names()

missing_temp_2 <- left_join(fish_fam, missing_temp_2)

missing_temp_2 <- missing_temp_2 %>%
  mutate(temp_preferred = as.numeric(temp_preferred)) %>%
  na.omit() %>%
  group_by(family) %>%
  summarise(temp_preferred = mean(temp_preferred)) %>%
  ungroup()

filled_data <- filled_data %>%
  left_join(missing_temp_2, by = "family") %>%
  mutate(temp_preferred = coalesce(temp_preferred.x, temp_preferred.y)) %>%
  dplyr::select(-temp_preferred.x, -temp_preferred.y)


# Fecundity Mean
missing_fec_2 <- fecundity() %>%
  filter(SpecCode %in% fish_fam$spec_code) %>%
  dplyr::select(SpecCode, FecundityMean) %>%
  janitor::clean_names()

missing_fec_2 <- left_join(fish_fam, missing_fec_2)

missing_fec_2 <- missing_fec_2 %>%
  mutate(fecundity_mean = as.numeric(fecundity_mean)) %>%
  na.omit() %>%
  group_by(family) %>%
  summarise(fecundity_mean = mean(fecundity_mean)) %>%
  ungroup()

filled_data <- filled_data %>%
  left_join(missing_fec_2, by = "family") %>%
  mutate(fecundity_mean = coalesce(fecundity_mean.x, fecundity_mean.y)) %>%
  dplyr::select(-fecundity_mean.x, -fecundity_mean.y)


# Oxygen Consumption
missing_ox_2 <- oxygen() %>%
  filter(SpecCode %in% fish_fam$spec_code) %>%
  dplyr::select(SpecCode, OxygenCons) %>%
  janitor::clean_names()

missing_ox_2 <- left_join(fish_fam, missing_ox_2)

missing_ox_2 <- missing_ox_2 %>%
  mutate(oxygen_cons = as.numeric(oxygen_cons)) %>%
  na.omit() %>%
  group_by(family) %>%
  summarise(oxygen_cons = mean(oxygen_cons)) %>%
  ungroup()

filled_data <- filled_data %>%
  left_join(missing_ox_2, by = "family") %>%
  mutate(oxygen_cons = coalesce(oxygen_cons.x, oxygen_cons.y)) %>%
  dplyr::select(-oxygen_cons.x, -oxygen_cons.y)


# Temp Range
missing_range_2 <- stocks() %>%
  filter(SpecCode %in% fish_fam$spec_code) %>%
  dplyr::select(SpecCode, TempMin, TempMax) %>%
  mutate(temp_range = TempMax - TempMin) %>%
  janitor::clean_names()

missing_range_2 <- left_join(fish_fam, missing_range_2)

missing_range_2 <- missing_range_2 %>%
  mutate(temp_range = as.numeric(temp_range)) %>%
  na.omit() %>%
  group_by(family) %>%
  summarise(temp_range = mean(temp_range)) %>%
  ungroup()

filled_data <- filled_data %>%
  left_join(missing_range_2, by = "family") %>%
  mutate(temp_range = coalesce(temp_range.x, temp_range.y)) %>%
  dplyr::select(-temp_range.x, -temp_range.y)


```

##### Write new csv including family filled variables

```{r}
write_csv(filled_data, here::here("01_data", "02_processed_data", "filled_data.csv"))
```

##### Assess missing data after family fills

```{r}
family_filled <- miss_var_summary(filled_data)

# New missing plot
filled_plot <- filled_data %>% 
  dplyr::select(
    oxygen_cons, temp_range, fecundity_mean, longevity_wild,
    temp_preferred, dev_place
  ) %>%
  rename(
    "Oxygen Consumption Rate" = oxygen_cons,
    "Temperature Range" = temp_range,
    "Average Fecundity" = fecundity_mean,
    "Maximum Recorded Lifespan" = longevity_wild,
    "Optimal Temperature" = temp_preferred,
    "Place of Development" = dev_place
  )

m3 <- gg_miss_var(filled_plot)  +
  theme_bw(base_size = 16) +
  labs(y = "Missing Data After Family-Level Fill", x = "") +
  theme(axis.title.x = element_text(vjust = -1))
family_filled_fig

```


```{r}
missing_fig <- (m1 / m2 / m3) +
  plot_annotation(tag_levels = "A") &
  theme_bw(base_size = 16)
missing_fig

ggsave(missing_fig, file = paste0("figS1.pdf"), path = here::here("03_figs"), height = 12, width = 8)
```
