==> 02_scripts/03_nmds.Rmd <==
---
title: "PCA"
output: html_document
date: "2023-04-13"
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE, message=FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)

library(tidyverse)
library(rfishbase)
library(FactoMineR)
library(factoextra)
library(car)
library(vegan)
library(patchwork)
library(cowplot)
library(ggpubr)
library(naniar)
library(pairwiseAdonis)
```

### Load data

```{r}
filled_data <- read_csv(here::here("01_data", "02_processed_data", "filled_data.csv")) %>%
  mutate_if(is.character, as.factor)
```

```{r}
scale_values <- function(x){(x-min(x, na.rm = TRUE))/(max(x, na.rm = TRUE)-min(x, na.rm = TRUE))}
```

```{r}
filled_data$temp_preferred = scale_values(filled_data$temp_preferred)
filled_data$longevity_wild =scale_values(filled_data$longevity_wild)
# filled_data$sp_richness = scale_values(filled_data$sp_richness)
filled_data$length = scale_values(filled_data$length)
filled_data$r_fin = scale_values(filled_data$r_fin)
filled_data$diet_troph3 = scale_values(filled_data$diet_troph3)
filled_data$temp_range =scale_values(filled_data$temp_range)
filled_data$oxygen_cons =scale_values(filled_data$oxygen_cons)
```

### FAMD

```{r}

famd_1 <- filled_data %>%
  dplyr::select(
    temp_preferred, fecundity_mean, longevity_wild, length,
    r_fin, diet_troph3, oxygen_cons, temp_range, body_shape_i,
    dev_place,  demers_pelag,
    class2
  ) %>%
  mutate(class2 = as.factor(class2)) %>%
  na.omit()


f1 <- FAMD(famd_1[, 1:11], graph = FALSE)
f1

eigen_f1 <- f1$eig
eigen_f1

# Contribution to the first dimension
c1 <- fviz_contrib(f1, "var", axes = 1) + scale_y_continuous(breaks = seq(0,25,by = 0.5))
# Contribution to the second dimension  
c2 <- fviz_contrib(f1, "var", axes = 2) + scale_y_continuous(breaks = seq(0,26,by = 0.5))
c1 + c2

f1_var = as.data.frame(f1$var) 

rownames(f1_var) <- c("Optimal Temperature", "Average Fecundity", "Lifespan", "Length", 
                      "Population Growth Rate", "Trophic Level", "Oxygen Consumption", "Temperature Range",
                      "Body Shape", "Place of Development","Habitat Type")

f1_var <- f1_var %>% 
  dplyr::select(coord.Dim.1, coord.Dim.2) %>% 
  rownames_to_column(var = "var") %>% 
  mutate(coord.Dim.1 = case_when(
    var == "Optimal Temperature" ~ -coord.Dim.1,
    var == "Oxygen Consumption" ~ -coord.Dim.1,
    var == "Population Growth Rate" ~ -coord.Dim.1,
    TRUE ~ coord.Dim.1
  )) %>% 
  mutate(coord.Dim.2 = case_when(
    var == "Temperature Range" ~ -coord.Dim.2,
    TRUE ~ coord.Dim.2
  ))


# PLOTS ARROWS
plot1 <- fviz_famd_var(f1, "quanti.var",
  repel = TRUE,
  col.var = "black"
) +
  theme_bw(base_size = 16) +
  labs(x = "Dimmension 1", y = "Dimmension 2", title = NULL) +
  scale_y_continuous(position = "right")


f1_fig <- fviz_famd_ind(f1, label = "none", habillage = famd_1$class2, palette = hcl.colors(6, "viridis"), 
                        addEllipses = TRUE, ellipse.level = 0.95, invisible = "quali.var", alpha.ind = 0.4, 
                        ellipse.alpha = 0.2) +
  theme_bw(base_size = 16) +
  labs(x = "FAMD1 (13.4%)", y = "FAMD2 (9.9%)", color = "Movement Group (Homerange/PLD)", 
       fill = "Movement Group (Homerange/PLD)", title = NULL) +
  scale_y_continuous(position = "right") +
  theme(legend.position = "left") + 
  geom_segment(data = f1_var, aes(x = 0, xend = coord.Dim.1*7, y = 0, yend = coord.Dim.2*7), 
               arrow = arrow(length = unit(0.25, "cm")), colour = "grey10", lwd = 0.5) +
  ggrepel::geom_label_repel(data = f1_var, aes(x = coord.Dim.1*7, y = coord.Dim.2*7, label = var), 
                            cex = 3, direction = "both", segment.size = 0.25) +
  geom_vline(xintercept = 0, linetype = "dashed", color = "grey") + 
  geom_hline(yintercept = 0, linetype = "dashed", color = "grey")

f1_fig

ggsave(f1_fig, file = paste0("figS2.pdf"), path = here::here("03_figs"), height = 8, width = 12)

```


##### NMDS: All significant traits

```{r}

com_data <- filled_data %>%
  dplyr::select(
    temp_preferred, longevity_wild, length, r_fin, diet_troph3, temp_range, oxygen_cons,
    species, order, class2, class3
  ) %>% 
  na.omit()

com <- com_data %>%
  dplyr::select(-species, -order, -class2, -class3)

m_com <- as.matrix(com)

set.seed(123)
nmds <- metaMDS(m_com, distance = "euclidean", autotransform = FALSE)
nmds

# extract NMDS scores (x and y coordinates)
filled_scores <- as.data.frame(scores(nmds)$sites)

# add columns to data frame
filled_scores$class2 <- com_data$class2
filled_scores$species <- com_data$species
filled_scores$order <- com_data$order

filled_scores <- filled_scores %>%
  mutate(class2 = as.factor(class2), order = as.factor(order), species = as.factor(species))

species_scores <- as.data.frame(nmds[["species"]])

rownames(species_scores) <- c("Optimal Temperature", "Lifespan",  "Length",
                               "Population Growth Rate", "Trophic Level", "Temperature Range", "Oxygen Consumption")

species_scores <- species_scores %>%
  rownames_to_column(var = "var")

centroid <- filled_scores %>%
  group_by(class2) %>%
  summarize(NMDS1 = mean(NMDS1), NMDS2 = mean(NMDS2))

nmds_1 <- ggplot(filled_scores, aes(x = NMDS1, y = NMDS2)) +
  geom_vline(xintercept = 0, linetype = "dashed", color = "grey") +
  geom_hline(yintercept = 0, linetype = "dashed", color = "grey") +
  stat_ellipse(geom = "polygon", aes(color = class2, fill = class2), linewidth = 0.25, alpha = 0.2, show.legend = FALSE) +
  geom_point(aes(color = class2), alpha = 0.3) +
  theme_bw(base_size = 16) +
  geom_point(data = centroid, aes(color = class2), size = 4, alpha = 0.9) +
  labs(x = "nMDS1", y = "nMDS2", color = "Movement Group (Homerange/PLD)") +
  scale_color_viridis_d() +
  scale_fill_viridis_d() +
  geom_segment(data = species_scores, aes(x = 0, xend = MDS1, y = 0, yend = MDS2), arrow = arrow(length = unit(0.25, "cm")), linewidth = 0.5, colour = "grey20") +
  ggrepel::geom_label_repel(data = species_scores, aes(x = MDS1, y = MDS2, label = var), fill = "white", cex = 3, direction = "both", segment.size = .25) +
  scale_y_continuous(position = "right") +
  theme(legend.position = "left", panel.grid.major = element_blank(), panel.grid.minor = element_blank()) +
  annotate("text", x = -0.5, y = 0.45, label = "Stress: 0.13")

nmds_1

ggsave(nmds_1, file = paste0("figS3.pdf"), path = here::here("03_figs"), height = 8, width = 12)


dst <- dist(com, method = "euclidean")

bd <- betadisper(dst, com_data$class2)

anova(bd)

permutest(bd, pairwise = TRUE, permutations = 9999)

adonis2(com_data$class3 ~ m_com, permutations = 9999)

pairwise.adonis(m_com, com_data$class3, perm = 9999)

```


##### NMDS 4 key traits: Optimal temperature, length, longevity, trophic level

```{r}

com_data2 <- filled_data %>%
  dplyr::select(
    temp_preferred, length, longevity_wild, r_fin,
    species, order, class2, class3
  ) %>%
  na.omit()


com2 <- com_data2 %>%
  dplyr::select(-species, -order, -class2, -class3)

m_com2 <- as.matrix(com2)

set.seed(123)
nmds2 <- metaMDS(m_com2, distance = "euclidean", autotransform = FALSE, wascores = TRUE)
nmds2

# extract NMDS scores (x and y coordinates)
filled_scores2 <- as.data.frame(scores(nmds2)$sites)

# add columns to data frame
filled_scores2$class2 <- com_data2$class2
filled_scores2$species <- com_data2$species
filled_scores2$order <- com_data2$order

filled_scores2 <- filled_scores2 %>%
  mutate(class2 = as.factor(class2), order = as.factor(order), species = as.factor(species))

species_scores2 <- as.data.frame(nmds2[["species"]])

rownames(species_scores2) <- c("Optimal Temperature", "Length", "Lifespan", "Population Growth Rate")

species_scores2 <- species_scores2 %>% 
  rownames_to_column(var = "var")

centroid2 <- filled_scores2 %>%
  group_by(class2) %>%
  summarize(NMDS1 = mean(NMDS1), NMDS2 = mean(NMDS2))

nmds_2 <- ggplot(filled_scores2, aes(x = NMDS1, y = NMDS2)) +
  geom_vline(xintercept = 0, linetype = "dashed", color = "grey") +
  geom_hline(yintercept = 0, linetype = "dashed", color = "grey") +
  stat_ellipse(geom = "polygon", aes(color = class2, fill = class2), linewidth = 0.25, alpha = 0.2, show.legend = FALSE) +
  geom_point(aes(color = class2), alpha = 0.3) +
  theme_bw(base_size = 22) +
  geom_point(data = centroid2, aes(color = class2), size = 4, alpha = 0.9) +
  labs(x = "nMDS1", y = "nMDS2", color = "Movement Group (home range/PLD)") +
  scale_color_viridis_d() +
  scale_fill_viridis_d() +
  geom_segment(data = species_scores2, aes(x = 0, xend = MDS1, y = 0, yend = MDS2), arrow = arrow(length = unit(0.25, "cm")), linewidth = 0.5, colour = "grey20") +
  ggrepel::geom_label_repel(data = species_scores2, aes(x = MDS1, y = MDS2, label = var), size = 5, fill = "white", cex = 3, direction = "both", segment.size = .25) +
  scale_y_continuous(position = "right") +
  theme(legend.position = "left", panel.grid.major = element_blank(), panel.grid.minor = element_blank()) +
  annotate("text", x = -0.5, y = 0.45, label = "Stress: 0.07")

nmds_2

ggsave(nmds_2, file = paste0("fig2.pdf"), path = here::here("03_figs"), height = 8, width = 15)


dst2 <- dist(com2, method = "euclidean")

bd2 <- betadisper(dst2, com_data2$class2)

anova(bd2)

permutest(bd2, pairwise = TRUE, permutations = 9999)

adonis2(com_data2$class3 ~ m_com2, permutations = 9999)

pairwise.adonis(m_com2, com_data2$class3, perm = 9999)

```
