==> 02_scripts/04_stats.Rmd <==
---
title: "Statistical Analyses"
author: "Lex Rosenberg"
date: "2023-06-06"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE, message=FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)

library(tidyverse)
library(here)
library(ggpubr)
library(effectsize)
library(rstatix)
library(FSA)
library(chisq.posthoc.test)
library(MASS)
library(sjPlot)
library(marginaleffects)
library(patchwork)
library(broom)
```

### Load data

```{r}
filled_data <- read_csv(here::here("01_data", "02_processed_data", "filled_data.csv")) %>%
  mutate_if(is.character, as.factor)
```

### Characterize data

```{r}
# Cohort sample size
sample_size <- filled_data %>%
  group_by(class2) %>%
  summarise(n = length(class2))

# Trait means, medians, and sd by cohort
trait_summary <- filled_data %>%
  dplyr::select(class2, length, r_fin, temp_preferred, longevity_wild) %>%
  group_by(class2) %>%
  summarise(across(
    where(is.numeric),
    list(
      mean = ~ mean(., na.rm = TRUE),
      sd = ~ sd(., na.rm = TRUE),
      median = ~ median(., na.rm = TRUE)
    )
  ))

trait_summary2 <- filled_data %>%
  dplyr::select(class2, diet_troph3, temp_range, oxygen_cons, fecundity_mean) %>%
  group_by(class2) %>%
  summarise(across(
    where(is.numeric),
    list(
      mean = ~ mean(., na.rm = TRUE),
      sd = ~ sd(., na.rm = TRUE),
      median = ~ median(., na.rm = TRUE)
    )
  ))
```

### Build linear models

```{r}
lm1 <- lm(log(home_range) ~ length * r_fin * longevity_wild * temp_preferred, data = filled_data)
lm2 <- lm(pld ~ length * r_fin * longevity_wild * temp_preferred, data = filled_data)

summary(lm1)
AIC(lm1)
Anova(lm1, type = "III")

summary(lm2)
AIC(lm2)
Anova(lm2, type = "III")

tidy_lm1 <- tidy(lm1)
tidy_lm2 <- tidy(lm2)

write_csv(tidy_lm1.1, here::here("01_data", "02_processed_data", "hr_lm.csv"))
write_csv(tidy_lm2.1, here::here("01_data", "02_processed_data", "pld_lm.csv"))

p1 <- plot_comparisons(lm1, variables = "length", condition = c("longevity_wild", "r_fin"), comparison = "difference") +
  labs(
    x = "Maximum Lifespan (yr)",
    y =  expression("Marginal Effect of Length (cm) on Home Range (km"^2*")"),
    color = "Growth Rate",
    fill = "Growth Rate",
  ) +
  theme_bw(base_size = 18) +
  scale_fill_viridis_d(labels = function(x) paste0(round(as.numeric(x), 2))) +
  scale_color_viridis_d(labels = function(x) paste0(round(as.numeric(x), 2))) 
  # xlim(NA, 150) +
  # ylim(-0.25, 0.15)

p2 <- plot_comparisons(lm2, variables = "length", condition = c("longevity_wild", "r_fin")) +
  labs(
    x = "Maximum Lifespan (yr)",
    y = "Marginal Effect of Length (cm) on PLD (days)",
    color = "Growth Rate",
    fill = "Growth Rate",
  ) +
  theme_bw(base_size = 18) +
  scale_fill_viridis_d(labels = function(x) {
    paste0(round(as.numeric(x), 2))
  }) +
  scale_color_viridis_d(labels = function(x) {
    paste0(round(as.numeric(x), 2))
  }) 

p3 <- plot_comparisons(lm2, variables = "length", condition = list("temp_preferred", "r_fin")) +
  labs(
    x = "Optimal Temperature (ºC)",
    y = "Marginal Effect of Length (cm) on PLD (days)", 
    color = "Growth Rate",
    fill = "Growth Rate"
  ) +
  theme_bw(base_size = 18) +
  scale_fill_viridis_d(labels = function(x) paste0(round(as.numeric(x), 2))) +
  scale_color_viridis_d(labels = function(x) paste0(round(as.numeric(x), 2))) 
  # xlim(NA, 45) +
  # ylim(-2, NA)

interact_plot <- (p3 + guide_area() ) / (p1 + p2) + plot_layout(guides = "collect") + plot_annotation(tag_levels = "A")

ggsave(interact_plot, file = paste0("fig5.pdf"), path = here::here("03_figs"), height = 12, width = 15)
```


### Variable Inflation Factors (VIF)

```{r}
model <- lm(
  class3 ~ temp_preferred + temp_max + fecundity_mean + longevity_wild +
    length + kfin + r_fin + diet_troph3,
  data = filled_data
)

summary(model)

vif(model)

# create vector of VIF values
vif_values <- vif(model)

# create horizontal bar chart to display each VIF value
barplot(vif_values, main = "VIF Values", horiz = TRUE, col = "steelblue", xlim = c(0, 6), las = 2)

# add vertical line at 5
abline(v = 5, lwd = 3, lty = 2)


# model2 <- lm(class2 ~ temp_preferred + length + r_fin, data = filled_data)
# summary(model)
```


### ANOVA/Kruskal-Wallis for Numeric/Continuous Variables

```{r}
# optimal temperature -----------------------------------------------

temp_data <- filled_data %>%
  filter(!is.na(temp_preferred)) %>%
  filter(temp_preferred > 0)

# look at outliers
outliers_temp <- boxplot.stats(temp_data$temp_preferred)$out
boxplot(temp_data$temp_preferred, main = "Preferred Temp", boxwex = 0.1)
mtext(paste("Outliers: ", paste(outliers_temp, collapse = ", ")), cex = 0.4)

# build the model
model_temp <- aov(temp_preferred ~ class2, data = temp_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_temp))
# not normal

# model with log transformed variable
model_temp_log <- aov(log(temp_preferred) ~ class2, data = temp_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_temp_log))
# more abnormal

# test for homogeneity of variance
plot(model_temp, 1)
# looks fine

# run Kruskal-Wallis test (non-parametric version of ANOVA)
kruskal.test(temp_preferred ~ class2, data = temp_data)
# chi-squared = 77.77, df = 5, p-value = 2.455e-15

# get effect size
temp_data %>% kruskal_effsize(temp_preferred ~ class2, ci = TRUE)
# eta2 = 0.124, MODERATE effect

# Kruskal-Wallis is significant, run Dunn's test
dunnTest(temp_preferred ~ class2, data = temp_data, method = "bh")



# maximum temperature -----------------------------------------------

temp_max_data <- filled_data %>%
  filter(!is.na(temp_max))

# look at outliers
outliers_temp <- boxplot.stats(temp_data$temp_max)$out
boxplot(temp_data$temp_max, main = "Max Temp", boxwex = 0.1)
mtext(paste("Outliers: ", paste(outliers_temp, collapse = ", ")), cex = 0.4)

# build the model
model_temp_max <- aov(temp_max ~ class2, data = temp_max_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_temp_max))
# not normal

# model with log transformed variable
model_temp_max_log <- aov(log(temp_max) ~ class2, data = temp_max_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_temp_max_log))
# more abnormal

# test for homogeneity of variance
plot(model_temp_max, 1)
# looks fine

# run Kruskal-Wallis test (non-parametric version of ANOVA)
kruskal.test(temp_max ~ class2, data = temp_max_data)
# chi-squared = 43.754, df = 5, p-value = 2.599e-08

# get effect size
temp_max_data %>% kruskal_effsize(temp_max ~ class2, ci = TRUE)
# eta2 = 0.178, LARGE effect

# Kruskal-Wallis is significant, run Dunn's test
dunnTest(temp_max ~ class2, data = temp_max_data, method = "bh")



# length -------------------------------------------------------

# check for outliers - there are a lot?
outliers_length <- boxplot.stats(filled_data$length)$out
boxplot(filled_data$length, main = "Length", boxwex = 0.1)
mtext(paste("Outliers: ", paste(outliers_length, collapse = ", ")), cex = 0.4)

# build the model
model_length <- aov(length ~ class2, data = filled_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_length))
# not normal

# model with log transformed length
model_length_log <- aov(log(length) ~ class2, data = filled_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_length_log))
# looks normal

# run Shapiro-Wilk test of normality
shapiro_test(residuals(model_length_log))
# p-value = 0.0669, normal

# test for homogeneity of variance
plot(model_length_log, 1)
# looks good

summary(model_length_log)
# p-value < 2e-16 ***

# effect size
effectsize::cohens_f(model_length_log)

effectsize::eta_squared(model_length_log)
interpret_eta_squared(0.18, rules = "cohen1992")
# eta2 = 0.18, LARGE effect

# post hoc Tukey test
TukeyHSD(model_length_log)



# trophic level --------------------------------------------------

# look at outliers
outliers_diet_troph3 <- boxplot.stats(filled_data$diet_troph3)$out
boxplot(filled_data$diet_troph3, main = "Trophic Position", boxwex = 0.1)
mtext(paste("Outliers: ", paste(outliers_diet_troph3, collapse = ", ")), cex = 0.4)

# build the model
model_diet_troph3 <- aov(diet_troph3 ~ class2, data = filled_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_diet_troph3))
# not normal

# model with transformed variable
model_diet_troph3_log <- aov(log(diet_troph3) ~ class2, data = filled_data)
model_diet_troph3_exp <- aov(exp(diet_troph3) ~ class2, data = filled_data)
model_diet_troph3_sqrt <- aov(sqrt(diet_troph3) ~ class2, data = filled_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_diet_troph3_log))
ggqqplot(residuals(model_diet_troph3_exp))
ggqqplot(residuals(model_diet_troph3_sqrt))
# none are normal

# test for homogeneity of variance
plot(model_diet_troph3, 1)
# looks good

# run Kruskal-Wallis test (non-parametric version of ANOVA)
kruskal.test(diet_troph3 ~ class2, data = filled_data)
# chi-squared = 36.497, df = 5, p-value = 7.555e-07

# get effect size
filled_data %>% kruskal_effsize(diet_troph3 ~ class2, ci = TRUE)
# eta2 = 0.0525, SMALL effect

# Kruskal-Wallis is significant, run Dunn's test
dunnTest(diet_troph3 ~ class2, data = filled_data, method = "bh")



# lifespan -----------------------------------------------

longevity_data <- filled_data %>%
  filter(!is.na(longevity_wild))

# look at outliers
outliers_long <- boxplot.stats(longevity_data$longevity_wild)$out
boxplot(longevity_data$longevity_wild, main = "Longevity in the Wild", boxwex = 0.1)
mtext(paste("Outliers: ", paste(outliers_long, collapse = ", ")), cex = 0.4)

# build the model
model_long <- aov(longevity_wild ~ class2, data = longevity_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_long))
# not normal

# model with log transformed variable
model_long_log <- aov(log(longevity_wild) ~ class2, data = longevity_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_long_log))
# almost normal?

# run Shapiro-Wilk test of normality
shapiro_test(residuals(model_long_log))
# p-value = 0.00000807, not normal

# test for homogeneity of variance
plot(model_long, 1)
# looks fine?

# run Kruskal-Wallis test (non-parametric version of ANOVA)
kruskal.test(longevity_wild ~ class2, data = longevity_data)
# chi-squared = 25.928, df = 5, p-value = 9.216e-05

# get effect size
longevity_data %>% kruskal_effsize(longevity_wild ~ class2, ci = TRUE)
# eta2 = 0.0384, SMALL effect

# Kruskal-Wallis is significant, run Dunn's test
dunnTest(longevity_wild ~ class2, data = longevity_data, method = "bh")

# r_fin --------------------------------------------------------

# look at outliers
outliers_r_fin <- boxplot.stats(filled_data$r_fin)$out
boxplot(filled_data$r_fin, main = "Growth Rate", boxwex = 0.1)
mtext(paste("Outliers: ", paste(outliers_r_fin, collapse = ", ")), cex = 0.4)

# build the model
model_r_fin <- aov(r_fin ~ class2, data = filled_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_r_fin))
# not normal

# model with log transformed variable
model_r_fin_log <- aov(log(r_fin) ~ class2, data = filled_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_r_fin_log))
# normal?

# run Shapiro-Wilk test of normality
shapiro_test(residuals(model_r_fin_log))
# p-value = 0.000508, not normal

# test for homogeneity of variance
plot(model_r_fin, 1)
# looks fine

# run Kruskal-Wallis test (non-parametric version of ANOVA)
kruskal.test(r_fin ~ class2, data = filled_data)
# chi-squared = 42.245, df = 5, p-value = 5.255e-08

# get effect size
filled_data %>% kruskal_effsize(r_fin ~ class2, ci = TRUE)
# eta2 = 0.0621, MODERATE effect

# Kruskal-Wallis is significant, run Dunn's test
dunnTest(r_fin ~ class2, data = filled_data, method = "bh")



# temp_range ---------------------------------------------------

temprange_data <- filled_data %>%
  filter(!is.na(temp_range))

# look at outliers
outliers_trange <- boxplot.stats(temprange_data$temp_range)$out
boxplot(temprange_data$temp_range, main = "Temperature Range", boxwex = 0.1)
mtext(paste("Outliers: ", paste(outliers_trange, collapse = ", ")), cex = 0.4)

# build the model
model_temprange <- aov(temp_range ~ class2, data = temprange_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_temprange))
# not normal

# model with log transformed variable
model_temprange_log <- aov(log(temp_range) ~ class2, data = temprange_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_temprange_log))
# almost normal?

# run Shapiro-Wilk test of normality
shapiro_test(residuals(model_temprange_log))
# p-value = 0.0000239, not normal

# test for homogeneity of variance
plot(model_temprange, 1)
# looks good

# run Kruskal-Wallis test (non-parametric version of ANOVA)
kruskal.test(temp_range ~ class2, data = temprange_data)
# chi-squared = 19.907, df = 5, p-value = 0.001301

# get effect size
filled_data %>% kruskal_effsize(temp_range ~ class2, ci = TRUE)
# eta2 = 0.0248, SMALL effect

# Kruskal-Wallis is significant, run Dunn's test
dunnTest(temp_range ~ class2, data = temprange_data, method = "bh")



# oxygen_cons --------------------------------------------------

oxygen_data <- filled_data %>%
  filter(!is.na(oxygen_cons))

# look at outliers
outliers_ox <- boxplot.stats(oxygen_data$oxygen_cons)$out
boxplot(oxygen_data$oxygen_cons, main = "Oxygen Consumption", boxwex = 0.1)
mtext(paste("Outliers: ", paste(outliers_ox, collapse = ", ")), cex = 0.4)

# build the model
model_oxcons <- aov(oxygen_cons ~ class2, data = oxygen_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_oxcons))
# not normal

# model with log transformed variable
model_oxcons_log <- aov(log(oxygen_cons) ~ class2, data = oxygen_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_oxcons_log))
# not normal

# test for homogeneity of variance
plot(model_oxcons, 1)
# not sure about this?

# run Kruskal-Wallis test (non-parametric version of ANOVA)
kruskal.test(oxygen_cons ~ class2, data = oxygen_data)
# chi-squared = 28.931, df = 5, p-value = 2.392e-05

# get effect size
filled_data %>% kruskal_effsize(oxygen_cons ~ class2, ci = TRUE)
# eta2 = 0.0399, SMALL effect

# Kruskal-Wallis is significant, run Dunn's test
dunnTest(oxygen_cons ~ class2, data = oxygen_data, method = "bh")



# fecundity_mean -----------------------------------------------

# look at outliers
outliers_fecundity <- boxplot.stats(filled_data$fecundity_mean)$out
boxplot(filled_data$fecundity_mean, main = "Fecundity", boxwex = 0.1)
mtext(paste("Outliers: ", paste(outliers_fecundity, collapse = ", ")), cex = 0.4)

# build the model
model_fecundity <- aov(fecundity_mean ~ class2, data = filled_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_fecundity))
# not normal

# model with log transformed variable
model_fecundity_log <- aov(log(fecundity_mean) ~ class2, data = filled_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_fecundity_log))
# not normal

# test for homogeneity of variance
plot(model_fecundity, 1)
# horizontal, looks good

# run Kruskal-Wallis test (non-parametric version of ANOVA)
kruskal.test(fecundity_mean ~ class2, data = filled_data)
# chi-squared = 42.357, df = 5, p-value = 4.988e-08

# get effect size
filled_data %>% kruskal_effsize(fecundity_mean ~ class2, ci = TRUE)
# eta2 = 0.0623, MODERATE effect

# Kruskal-Wallis is significant, run Dunn's test
dunnTest(fecundity_mean ~ class2, data = filled_data, method = "bh")
```



### Chi-Squared for Categorical Variables

```{r}
# demers_pelag -------------------------------------------------

# Pearson's Chi-squared test
chisq.test(filled_data$class2, filled_data$demers_pelag)
# X-squared = 232.88, df = 30, p-value < 2.2e-16 - there is a significant relationship

# contingency table
demers_table <- table(filled_data$class2, filled_data$demers_pelag)
demers_table

# post-hoc test
chisq.posthoc.test(demers_table, method = "bonferroni")


# movement_keyword ---------------------------------------------

# Pearson's Chi-squared test
chisq.test(filled_data$class2, filled_data$movement_keyword)
# X-squared = 309.78, df = 50, p-value < 2.2e-16

# contingency table
movement_table <- table(filled_data$class2, filled_data$movement_keyword)
movement_table

# post-hoc test
chisq.posthoc.test(movement_table, method = "bonferroni")


# body_shape_i -------------------------------------------------

# Pearson's Chi-squared test
chisq.test(filled_data$class2, filled_data$body_shape_i)
# X-squared = 28.07, df = 15, p-value = 0.02114

# contingency table
body_table <- table(filled_data$class2, filled_data$body_shape_i)
body_table

# post-hoc test
chisq.posthoc.test(body_table, method = "bonferroni")


# climate ------------------------------------------------------

# Pearson's Chi-squared test
chisq.test(filled_data$class2, filled_data$climate)
# X-squared = 17.694, df = 5, p-value = 0.003355

# contingency table
climate_table <- table(filled_data$class2, filled_data$climate)
climate_table

# post-hoc test
chisq.posthoc.test(climate_table, method = "bonferroni")


# dev_place ----------------------------------------------------

# Pearson's Chi-squared test
chisq.test(filled_data$class2, filled_data$dev_place)
# X-squared = 12.4531, df = 5, p-value = 0.0291

# contingency table
dev_table <- table(filled_data$class2, filled_data$dev_place)
dev_table

# post-hoc test
chisq.posthoc.test(dev_table, method = "bonferroni")


# food_mode ----------------------------------------------------

# Pearson's Chi-squared test
chisq.test(filled_data$class2, filled_data$food_mode)
# X-squared = 8.4677, df = 5, p-value = 0.1323
```
