==> 02_scripts/05_figures.Rmd <==
---
title: "Figures"
author: "Lex Rosenberg"
date: "2023-07-26"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE, message=FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)

library(tidyverse)
library(patchwork)
library(cowplot)
library(ggpubr)
library(ggrepel)
library(naniar)
```


### Load data

```{r}
filled_data <- read_csv(here::here("01_data", "02_processed_data", "filled_data.csv")) %>%
  mutate_if(is.character, as.factor)
```


### Summary figure

```{r}

cohort <- c("1: HR-mid/PLD-low vs 2: HR-low/PLD-mid", "1: HR-mid/PLD-low vs 3: HR-mid/PLD-high", "1: HR-mid/PLD-low vs 4: HR-mid/PLD-very high", "1: HR-mid/PLD-low vs 5: HR-mid/PLD-mid", "1: HR-mid/PLD-low vs 6: HR-high/PLD-mid",
            "2: HR-low/PLD-mid vs 3: HR-mid/PLD-high", "2: HR-low/PLD-mid vs 4: HR-mid/PLD-very high", "2: HR-low/PLD-mid vs 5: HR-mid/PLD-mid", "2: HR-low/PLD-mid vs 6: HR-high/PLD-mid",
            "3: HR-mid/PLD-high vs 4: HR-mid/PLD-very high", "3: HR-mid/PLD-high vs 5: HR-mid/PLD-mid", "3: HR-mid/PLD-high vs 6: HR-high/PLD-mid",
            "4: HR-mid/PLD-very high vs 5: HR-mid/PLD-mid", "4: HR-mid/PLD-very high vs 6: HR-high/PLD-mid",
            "5: HR-mid/PLD-mid vs 6: HR-high/PLD-mid")
PERMDISP <- c(0.001, 0.032, 0.771, 0.550, 0.587, 
              0.425, 0.17, 0.0006, 0.028, 
              0.529, 0.159, 0.305, 
              0.971, 0.98, 
              0.928)
PERMANOVA <- c(0.06, 0.002, 0.002, 0.002, 0.002, 
               0.002, 0.026, 0.002, 0.002,
               1, 0.002, 0.002, 
               0.215, 0.002,
               0.002)
Optimal_Temp <- c(0.01, 0.01, 0.01, 0.01, 0.01, 
                  0.01, 0.01, 0.01, 0.01,
                  0.430, 0.01, 0.01, 
                  0.01, 0.01,
                  0.1)
temp_arrows <- c("increase", "increase", "increase", "decrease", "decrease",
                 "increase", "increase","decrease", "decrease",
                 "increase","decrease", "decrease",
                 "decrease", "decrease",
                 "decrease")
Length <- c(0.9, 0.9, 0.9, 0.01, 0.01,
            0.9, 0.9, 0.01, 0.01, 
            0.9, 0.0007, 0.01, 
            0.34, 0.003,
            0.027)
length_arrows <- c("decrease", "decrease", "decrease", "decrease", "decrease",
                   "decrease", "decrease", "decrease", "decrease",
                   "decrease", "decrease", "decrease",
                   "decrease", "decrease",
                   "decrease")
Lifespan <- c(0.6, 0.0001, 0.6, 0.98, 0.25,
              0.00001, 0.61, 0.71, 0.14,
              0.388, 0.0006, 0.00001,
              0.57, 0.26,
              0.27)
life_arrows <- c("decrease", "decrease", "decrease", "no change", "increase",
                 "decrease", "decrease", "increase", "increase",
                 "increase", "increase", "increase",
                 "increase", "increase",
                 "increase")
Growth_Rate <- c(0.237, 0.0001, 0.258, 0.396, 0.117,
                 0.0001, 0.303, 0.959, 0.287,
                 0.322, 0.0001, 0.039,
                 0.303, 0.774,
                 0.321)
growth_arrows <- c("increase", "increase", "increase", "increase", "decrease",
                    "increase", "increase", "decrease", "increase",
                    "decrease", "decrease", "decrease",
                    "decrease", "decrease",
                    "increase")

# Create data frame
data <- data.frame(cohort, PERMDISP, PERMANOVA, Optimal_Temp, Length, Lifespan, Growth_Rate)

# Pivot to long data format
data_long <- data %>% 
  mutate(across(c("Optimal_Temp", "Length", "Lifespan", "Growth_Rate"), round, 4)) %>% 
  pivot_longer(-cohort) %>% 
  set_names(c("cohort", "variable", "p_value")) %>% 
  mutate(significant = case_when(
    p_value >= 0.05 ~ "no",
    p_value < 0.05 ~ "yes"
  ))

# Add arrow information
data1 <- data.frame(cohort, temp_arrows, length_arrows, life_arrows, growth_arrows) %>% 
  rename(Optimal_Temp = temp_arrows, Length = length_arrows, Lifespan = life_arrows, Growth_Rate = growth_arrows) %>% 
  pivot_longer(-cohort) %>% 
  set_names(c("cohort", "variable", "direction")) %>% 
  full_join(data_long, by = c("cohort","variable")) %>% 
  mutate(direction = fct_relevel(direction, c("increase", "no change", "decrease")))

# Reorder and rename levels
data1$variable <- factor(data1$variable, levels = c("PERMANOVA", "PERMDISP", "Optimal_Temp", "Length", "Lifespan", "Growth_Rate"))
data1$variable <- recode_factor(data1$variable, PERMANOVA = "PERMANOVA", PERMDISP = "PERMDISP", Optimal_Temp = "Optimal Temperature", Length = "Length", Lifespan = "Lifespan", Growth_Rate = "Growth Rate")



# Plot 
g <- ggplot(data1, aes(x = variable, y = cohort, fill = direction)) +
  geom_tile(color = "white", lwd = 1) +
  geom_point(data = data1 %>% filter(significant == "yes"), shape = 8, size = 3, color = "black", show.legend = FALSE) +
  theme_minimal(base_size = 12) +
  theme(axis.title = element_blank(),
        axis.text = element_text(face = "bold", color = "black"),
        axis.ticks = element_blank(),
        panel.grid.major = element_blank()) +
  labs(fill = "Direction of Effect") +
  scale_x_discrete(position = "top") +
  scale_y_discrete(limits = rev) +
  scale_fill_manual(values = c("#fde725", "#21918c", "#7568a0"), na.value = "darkgray")

g

ggsave(g, file = paste0("fig3.pdf"), path = here::here("03_figs"), height = 12, width = 12)


```


### Homerange vs PLD and current classes (Figure 1)

```{r}
# homerange vs PLD
fishbase_data_gmm <- filled_data %>%
  mutate(class2 = as.factor(class2))

hr_hist <- ggplot(fishbase_data_gmm, aes(log(home_range))) +
  geom_histogram() +
  theme_bw(base_size = 16) +
  labs(
    y = "Count",
    x = "ln(Predicted Home Range)",
    tag = "A"
  ) +
  ggtitle("Individual Distributions of Species Movement Extents")

pld_hist <- ggplot(fishbase_data_gmm, aes(pld)) +
  geom_histogram() +
  theme_bw(base_size = 16) +
  labs(
    y = "Count",
    x = "Predicted PLD"
  )

dist_plot <- (hr_hist / pld_hist)

p1 <- ggplot(fishbase_data_gmm, aes(log(home_range), pld)) +
  geom_jitter() +
  theme_bw(base_size = 16) +
  labs(
    y = "Predicted PLD",
    x = "ln(Predicted Home Range)",
    tag = "B"
  ) +
  ggtitle("Combined Distribution of Species Movement Extents")

p1


gmm_new_fig <- ggplot(fishbase_data_gmm, aes(magnitude_homerange, month_pld)) +
  theme_bw(base_size = 16) +
  labs(
    x = "Predicted Home Range Magnitude",
    y = "Predicted PLD Month",
    color = "Movement Group (Homerange/PLD)",
    tag = "C"
  ) +
  ylim(-0.1,6) +
  theme(legend.position = "bottom") +
  scale_color_viridis_d() +
  geom_label_repel(
    data = subset(fishbase_data_gmm, fishbase_name %in% c("Hogfish", "Yellow snapper", "Blackfin tuna")),
    aes(label = fishbase_name, color = class2),
    vjust = "outward", hjust = "outward",
    min.segment.length = 30,
    nudge_y = -1,
    show.legend = FALSE
  ) +
  geom_label_repel(
    data = subset(fishbase_data_gmm, fishbase_name %in% c("Chocolate hind", "Lingcod", "Nassau grouper")),
    aes(label = fishbase_name, color = class2),
    hjust = "inward",
    min.segment.length = 30,
    box.padding = 0.3,
    nudge_x = -2, #nudge_y = -0.8,
    show.legend = FALSE
  ) +
  geom_label_repel(
    data = subset(fishbase_data_gmm, fishbase_name %in% c("Lemon sole", "North Pacific hake", "Yellowstripe goatfish")),
    aes(label = fishbase_name, color = class2),
    vjust = "outward", hjust = "outward",
    min.segment.length = 30,
    nudge_x = -1, nudge_y = 0.8,
    show.legend = FALSE
  ) +
  geom_label_repel(
    data = subset(fishbase_data_gmm, fishbase_name %in% c("Atlantic herring", "Atlantic cod", "Summer flounder")),
    aes(label = fishbase_name, color = class2),
    #vjust = "outward", hjust = "outward",
    min.segment.length = 30,
    nudge_y = 0,
    nudge_x = 1,
    show.legend = FALSE
  ) +
  geom_label_repel(
    data = subset(fishbase_data_gmm, fishbase_name %in% c("Great barracuda", "Giant trevally", "Black seabass")),
    aes(label = fishbase_name, color = class2),
    vjust = "outward", hjust = "outward",
    min.segment.length = 30,
    nudge_y = 1, nudge_x = 0.3,
    show.legend = FALSE
  ) +
  geom_label_repel(
    data = subset(fishbase_data_gmm, fishbase_name %in% c("Common snook", "Yellowfin tuna", "Swordfish")),
    aes(label = fishbase_name, color = clas2),
    color = "goldenrod1",
    vjust = "outward", hjust = "outward",
    min.segment.length = 30,
    nudge_y = -0.5,
    show.legend = FALSE
  ) +
  geom_jitter(aes(color = class2)) +
  ggtitle("Distribution of Scaled Species Movement Extents")

gmm_new_fig
#ggsave(gmm_new_fig, file = paste0("gmm_fig.png"), path = here::here("04_figs"), height = 8, width = 15)

# panel
gmm_panel <- ((dist_plot) | p1) / gmm_new_fig &
  theme(plot.tag = element_text(face = "bold"))

gmm_panel
ggsave(gmm_panel, file = paste0("fig1.pdf"), path = here::here("03_figs"), height = 15, width = 15)

```


### Individual Plots

```{r}
# continuous

plot_data <- filled_data %>%
  mutate(
    class2 = as.factor(class2),
    class3 = as.factor(class3),
    order = as.factor(order)
  )

response <- "class2"
expl_cont <- c(
  "temp_preferred", "fecundity_mean",
  "longevity_wild", "sp_richness", "length",
  "r_fin", "diet_troph3", "oxygen_cons", "temp_range"
)

box_data <- function(x, y) {
  ggplot(plot_data, aes(x = .data[[y]], y = .data[[x]])) +
    geom_boxplot(aes(color = .data[[y]])) +
    theme_bw() +
    coord_flip() +
    labs(
      x = "Movement Group",
      color = "Class (Homerange/PLD)"
    ) +
    scale_color_viridis_d()
}

```

```{r}

# significant panel
temp_plot <- box_data(x = "temp_preferred", y = "class2") +
  labs(y = "Optimal Temperature (ºC)", x = "") +
  scale_x_discrete(position = "top") +
  theme(axis.text.y = element_blank())
length_plot <- box_data(x = "length", y = "class2") +
  labs(y = "Length (cm)", x = "") +
  scale_x_discrete(position = "top")
long_plot <- box_data(x = "longevity_wild", y = "class2") +
  labs(y = "Maximum Lifespan (yr)", x = "") +
  scale_x_discrete(position = "top")
diet_plot <- box_data(x = "diet_troph3", y = "class2") + 
  labs(y = "Trophic Position", x = "") +
  scale_x_discrete(position = "top") +
  theme(axis.text.y = element_blank())


r_fin_plot <- box_data(x = "r_fin", y = "class2") + 
  labs(y = "Growth Rate", x = "") +
  theme(axis.text.y = element_blank()) +
  scale_x_discrete(position = "top")
oxygen_plot <- box_data(x = "oxygen_cons", y = "class2") + 
  stat_kruskal_test(label.y = 2000, label.x = 5.8, position = "jitter") +
  labs(y = "Oxygen Consumption", x = "") +
  theme(axis.text.y = element_blank()) +
  scale_x_discrete(position = "top")
temp_range_plot <- box_data(x = "temp_range", y = "class2") + 
  stat_kruskal_test(label.y = 20, label.x = 3.5, position = "jitter") +
  labs(y = "Temperature Range", x = "") + 
  scale_x_discrete(position = "top")
fecund_plot <- box_data(x = "fecundity_mean", y = "class2") + 
  stat_kruskal_test(label.y = 40000000, label.x = 5.5, position = "jitter") +
  labs(y = "Fecundity", x = "") +
  scale_x_discrete(position = "top")


num_panel <- diet_plot + temp_range_plot + oxygen_plot +  fecund_plot +
  plot_layout(ncol = 2, guides = "collect") + 
  plot_annotation(tag_levels = "A") &
  theme(text = element_text(size = 14), legend.position = "none")

numeric_panel <- wrap_elements(num_panel) +
  labs(tag = "Movement Group") +
  theme(
    plot.tag = element_text(size = 14, angle = 270),
    plot.tag.position = "right"
  )

numeric_panel
ggsave(numeric_panel, file = paste0("figS4.pdf"), path = here::here("03_figs"), height = 11, width = 11)


ind_panel <- ((temp_plot | length_plot) / (r_fin_plot | long_plot)) +
  plot_layout(guides = "collect") +
  plot_annotation(tag_levels = "A") &
  theme(text = element_text(size = 14), legend.position = "none")

individual_panel <- wrap_elements(ind_panel) +
  labs(tag = "Movement Group") +
  theme(
    plot.tag = element_text(size = 14, angle = 270),
    plot.tag.position = "right"
  )

individual_panel

ggsave(individual_panel, file = paste0("fig4.pdf"), path = here::here("03_figs"), height = 8, width = 13)

```


```{r}
# categorical

expl_cat <- c(
  "body_shape_i", "dev_place", "demers_pelag", "iucn_category"
)

scatter_data <- function(x, y) {
  ggplot(plot_data, aes(x = .data[[x]], y = .data[[y]])) +
    geom_jitter(aes(color = .data[[y]])) +
    theme_bw() +
    labs(
      y = "Movement Group",
      color = "Class (homerange/PLD)"
    ) +
    scale_color_viridis_d() +
    theme(axis.text.x = element_text(angle = 40, hjust = 1))
}

```

```{r}

body_plot <- scatter_data(x = "body_shape_i", y = "class2") + 
  labs(y = "", x = "Body Shape") +
  scale_y_discrete(position = "right")
dev_plot <- scatter_data(x = "dev_place", y = "class2") + 
  labs(y = "", x = "Place of Development") +
  scale_y_discrete(position = "right")
demers_plot <- scatter_data(x = "demers_pelag", y = "class2") + 
  labs(y = "", x = "Habitat Type") +
  theme(axis.text.y = element_blank())

cat_panel <- demers_plot + dev_plot +  body_plot +
  plot_layout(ncol = 2, guides = "collect") + 
  plot_annotation(tag_levels = "A") &
  theme(text = element_text(size = 14), legend.position = "none")

categorical_panel <- wrap_elements(cat_panel) +
  labs(tag = "Movement Group") +
  theme(
    plot.tag = element_text(size = 14, angle = 270),
    plot.tag.position = "right"
  )

categorical_panel
ggsave(categorical_panel, file = paste0("figS5.pdf"), path = here::here("03_figs"), height = 11, width = 11)

```
