==> 02_scripts/06_sensitvity.Rmd <==
---
title: "06_sensitivity"
author: "Allie Caughman"
date: "2025-06-27"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)

library(tidyverse)
library(rfishbase)
library(FactoMineR)
library(factoextra)
library(car)
library(vegan)
library(patchwork)
library(cowplot)
library(ggpubr)
library(naniar)
library(pairwiseAdonis)
library(broom)
library(effectsize)
library(rstatix)
library(FSA)
library(chisq.posthoc.test)
library(MASS)
library(sjPlot)
```

##### Constrained Data

```{r}
filled_data <- read_csv(here::here("01_data", "02_processed_data", "filled_data.csv")) %>%
  mutate_if(is.character, as.factor) %>% 
  mutate(range_hr = floor(log10(pred_75)) - floor(log10((pred_25))),
         range_pld = ceiling(pred_75_pld/30) - ceiling(pred_25_pld/30)) %>% 
  mutate(keep_hr = case_when(
    range_hr <= 1 ~ "keep",
    !is.na(observed_homerange) ~ "keep"
  )) %>% 
  mutate(keep_pld = case_when(
    range_pld <= 1 ~ "keep",
    !is.na(observed_pld) ~ "keep"
  )) %>% 
  filter(!is.na(keep_hr)) %>% 
  filter(!is.na(keep_pld))
```

```{r}
scale_values <- function(x){(x-min(x, na.rm = TRUE))/(max(x, na.rm = TRUE)-min(x, na.rm = TRUE))}
```

```{r}
filled_data$temp_preferred = scale_values(filled_data$temp_preferred)
filled_data$longevity_wild =scale_values(filled_data$longevity_wild)
# filled_data$sp_richness = scale_values(filled_data$sp_richness)
filled_data$length = scale_values(filled_data$length)
filled_data$r_fin = scale_values(filled_data$r_fin)
filled_data$diet_troph3 = scale_values(filled_data$diet_troph3)
filled_data$temp_range =scale_values(filled_data$temp_range)
filled_data$oxygen_cons =scale_values(filled_data$oxygen_cons)
```


```{r}

com_data2 <- filled_data %>%
  dplyr::select(
    temp_preferred, length, longevity_wild, r_fin,
    species, order, class2, class3
  ) %>%
  na.omit()


com2 <- com_data2 %>%
  dplyr::select(-species, -order, -class2, -class3)

m_com2 <- as.matrix(com2)

set.seed(123)
nmds2 <- metaMDS(m_com2, distance = "euclidean", autotransform = FALSE, wascores = TRUE)
nmds2

# extract NMDS scores (x and y coordinates)
filled_scores2 <- as.data.frame(scores(nmds2)$sites)

# add columns to data frame
filled_scores2$class2 <- com_data2$class2
filled_scores2$species <- com_data2$species
filled_scores2$order <- com_data2$order

filled_scores2 <- filled_scores2 %>%
  mutate(class2 = as.factor(class2), order = as.factor(order), species = as.factor(species))

species_scores2 <- as.data.frame(nmds2[["species"]])

rownames(species_scores2) <- c("Optimal Temperature", "Length", "Lifespan", "Population Growth Rate")

species_scores2 <- species_scores2 %>% 
  rownames_to_column(var = "var")

centroid2 <- filled_scores2 %>%
  group_by(class2) %>%
  summarize(NMDS1 = mean(NMDS1), NMDS2 = mean(NMDS2))

nmds_2 <- ggplot(filled_scores2, aes(x = NMDS1, y = NMDS2)) +
  geom_vline(xintercept = 0, linetype = "dashed", color = "grey") +
  geom_hline(yintercept = 0, linetype = "dashed", color = "grey") +
  stat_ellipse(geom = "polygon", aes(color = class2, fill = class2), linewidth = 0.25, alpha = 0.2, show.legend = FALSE) +
  geom_point(aes(color = class2), alpha = 0.3) +
  theme_bw(base_size = 22) +
  geom_point(data = centroid2, aes(color = class2), size = 4, alpha = 0.9) +
  labs(x = "nMDS1", y = "nMDS2", color = "Movement Group (home range/PLD)") +
  scale_color_viridis_d() +
  scale_fill_viridis_d() +
  geom_segment(data = species_scores2, aes(x = 0, xend = MDS1, y = 0, yend = MDS2), arrow = arrow(length = unit(0.25, "cm")), linewidth = 0.5, colour = "grey20") +
  ggrepel::geom_label_repel(data = species_scores2, aes(x = MDS1, y = MDS2, label = var), size = 5, fill = "white", cex = 3, direction = "both", segment.size = .25) +
  scale_y_continuous(position = "right") +
  theme(legend.position = "left", panel.grid.major = element_blank(), panel.grid.minor = element_blank())

p1 = nmds_2

dst2 <- dist(com2, method = "euclidean")

bd2 <- betadisper(dst2, com_data2$class2)

anova(bd2)

permutest(bd2, pairwise = TRUE, permutations = 9999)

adonis2(com_data2$class3 ~ m_com2, permutations = 9999)

pairwise.adonis(m_com2, com_data2$class3, perm = 9999)
```

```{r}
filled_data <- read_csv(here::here("01_data", "02_processed_data", "filled_data.csv")) %>%
  mutate_if(is.character, as.factor) %>% 
  mutate(range_hr = floor(log10(pred_75)) - floor(log10((pred_25))),
         range_pld = ceiling(pred_75_pld/30) - ceiling(pred_25_pld/30)) %>% 
  mutate(keep_hr = case_when(
    range_hr <= 1 ~ "keep",
    !is.na(observed_homerange) ~ "keep"
  )) %>% 
  mutate(keep_pld = case_when(
    range_pld <= 1 ~ "keep",
    !is.na(observed_pld) ~ "keep"
  )) %>% 
  filter(!is.na(keep_hr)) %>% 
  filter(!is.na(keep_pld))
```

```{r}
lm1 <- lm(log(home_range) ~ length * r_fin * longevity_wild * temp_preferred, data = filled_data)
lm2 <- lm(pld ~ length * r_fin * longevity_wild * temp_preferred, data = filled_data)

summary(lm1)
AIC(lm1)
Anova(lm1, type = "III")

summary(lm2)
AIC(lm2)
Anova(lm2, type = "III")

tidy_lm1 <- tidy(lm1)
tidy_lm2 <- tidy(lm2)

write_csv(tidy_lm1, here::here("01_data", "02_processed_data", "hr_lm_constrained.csv"))
write_csv(tidy_lm2, here::here("01_data", "02_processed_data", "pld_lm_constrained.csv"))
```


```{r}
# optimal temperature -----------------------------------------------

temp_data <- filled_data %>%
  filter(!is.na(temp_preferred)) %>%
  filter(temp_preferred > 0)

temp_data %>% 
  group_by(class2) %>% 
  summarize(median(temp_preferred))

# look at outliers
outliers_temp <- boxplot.stats(temp_data$temp_preferred)$out
boxplot(temp_data$temp_preferred, main = "Preferred Temp", boxwex = 0.1)
mtext(paste("Outliers: ", paste(outliers_temp, collapse = ", ")), cex = 0.4)

# build the model
model_temp <- aov(temp_preferred ~ class2, data = temp_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_temp))
# not normal

# model with log transformed variable
model_temp_log <- aov(log(temp_preferred) ~ class2, data = temp_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_temp_log))
# more abnormal

# test for homogeneity of variance
plot(model_temp, 1)
# looks fine

# run Kruskal-Wallis test (non-parametric version of ANOVA)
kruskal.test(temp_preferred ~ class2, data = temp_data)
# chi-squared = 77.77, df = 5, p-value = 2.455e-15

# get effect size
temp_data %>% kruskal_effsize(temp_preferred ~ class2, ci = TRUE)
# eta2 = 0.124, MODERATE effect

# Kruskal-Wallis is significant, run Dunn's test
dunnTest(temp_preferred ~ class2, data = temp_data, method = "bh")

# length -------------------------------------------------------

# check for outliers - there are a lot?
outliers_length <- boxplot.stats(filled_data$length)$out
boxplot(filled_data$length, main = "Length", boxwex = 0.1)
mtext(paste("Outliers: ", paste(outliers_length, collapse = ", ")), cex = 0.4)

filled_data %>% 
  group_by(class2) %>% 
  summarize(mean = mean(length))

# build the model
model_length <- aov(length ~ class2, data = filled_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_length))
# not normal

# model with log transformed length
model_length_log <- aov(log(length) ~ class2, data = filled_data %>% filter(length > 0))

# create a QQ plot of residuals
ggqqplot(residuals(model_length_log))
# looks normal

# run Shapiro-Wilk test of normality
shapiro_test(residuals(model_length_log))
# p-value = 0.0669, normal

# test for homogeneity of variance
plot(model_length_log, 1)
# looks good

summary(model_length_log)
# p-value < 2e-16 ***

# effect size
effectsize::cohens_f(model_length_log)

effectsize::eta_squared(model_length_log)
interpret_eta_squared(0.53, rules = "cohen1992")
# eta2 = 0.18, LARGE effect

# post hoc Tukey test
TukeyHSD(model_length_log)

# lifespan -----------------------------------------------

longevity_data <- filled_data %>%
  filter(!is.na(longevity_wild))

longevity_data %>% 
  group_by(class2) %>% 
  summarise(median(longevity_wild))

# look at outliers
outliers_long <- boxplot.stats(longevity_data$longevity_wild)$out
boxplot(longevity_data$longevity_wild, main = "Longevity in the Wild", boxwex = 0.1)
mtext(paste("Outliers: ", paste(outliers_long, collapse = ", ")), cex = 0.4)

# build the model
model_long <- aov(longevity_wild ~ class2, data = longevity_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_long))
# not normal

# model with log transformed variable
model_long_log <- aov(log(longevity_wild) ~ class2, data = longevity_data %>% filter(longevity_wild > 0))

# create a QQ plot of residuals
ggqqplot(residuals(model_long_log))
# almost normal?

# run Shapiro-Wilk test of normality
shapiro_test(residuals(model_long_log))
# p-value = 0.00000807, not normal

# test for homogeneity of variance
plot(model_long, 1)
# looks fine?

# run Kruskal-Wallis test (non-parametric version of ANOVA)
kruskal.test(longevity_wild ~ class2, data = longevity_data)
# chi-squared = 25.928, df = 5, p-value = 9.216e-05

# get effect size
longevity_data %>% kruskal_effsize(longevity_wild ~ class2, ci = TRUE)
# eta2 = 0.0384, SMALL effect

# Kruskal-Wallis is significant, run Dunn's test
dunnTest(longevity_wild ~ class2, data = longevity_data, method = "bh")

# r_fin --------------------------------------------------------

# look at outliers
outliers_r_fin <- boxplot.stats(filled_data$r_fin)$out
boxplot(filled_data$r_fin, main = "Growth Rate", boxwex = 0.1)
mtext(paste("Outliers: ", paste(outliers_r_fin, collapse = ", ")), cex = 0.4)

filled_data %>% 
  group_by(class2) %>% 
  summarize(median(r_fin))

# build the model
model_r_fin <- aov(r_fin ~ class2, data = filled_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_r_fin))
# not normal

# model with log transformed variable
model_r_fin_log <- aov(log(r_fin) ~ class2, data = filled_data %>% filter(r_fin > 0))

# create a QQ plot of residuals
ggqqplot(residuals(model_r_fin_log))
# normal?

# run Shapiro-Wilk test of normality
shapiro_test(residuals(model_r_fin_log))
# p-value = 0.000508, not normal

# test for homogeneity of variance
plot(model_r_fin, 1)
# looks fine

# run Kruskal-Wallis test (non-parametric version of ANOVA)
kruskal.test(r_fin ~ class2, data = filled_data)
# chi-squared = 42.245, df = 5, p-value = 5.255e-08

# get effect size
filled_data %>% kruskal_effsize(r_fin ~ class2, ci = TRUE)
# eta2 = 0.0621, MODERATE effect

# Kruskal-Wallis is significant, run Dunn's test
dunnTest(r_fin ~ class2, data = filled_data, method = "bh")

```

##### Empirical Data

```{r}
filled_data_save <- read_csv(here::here("01_data", "02_processed_data", "filled_data.csv")) %>%
  mutate_if(is.character, as.factor)

filled_data = filled_data_save %>% 
  filter(!is.na(observed_homerange)) %>% 
  filter(!is.na(observed_pld))
```


```{r}
filled_data$temp_preferred = scale_values(filled_data$temp_preferred)
filled_data$longevity_wild =scale_values(filled_data$longevity_wild)
# filled_data$sp_richness = scale_values(filled_data$sp_richness)
filled_data$length = scale_values(filled_data$length)
filled_data$r_fin = scale_values(filled_data$r_fin)
filled_data$diet_troph3 = scale_values(filled_data$diet_troph3)
filled_data$temp_range =scale_values(filled_data$temp_range)
filled_data$oxygen_cons =scale_values(filled_data$oxygen_cons)
```


```{r}

com_data2 <- filled_data %>%
  dplyr::select(
    temp_preferred, length, longevity_wild, r_fin,
    species, order, class2, class3
  ) %>%
  na.omit()


com2 <- com_data2 %>%
  dplyr::select(-species, -order, -class2, -class3)

m_com2 <- as.matrix(com2)

set.seed(123)
nmds2 <- metaMDS(m_com2, distance = "euclidean", autotransform = FALSE, wascores = TRUE)
nmds2

# extract NMDS scores (x and y coordinates)
filled_scores2 <- as.data.frame(scores(nmds2)$sites)

# add columns to data frame
filled_scores2$class2 <- com_data2$class2
filled_scores2$species <- com_data2$species
filled_scores2$order <- com_data2$order

filled_scores2 <- filled_scores2 %>%
  mutate(class2 = as.factor(class2), order = as.factor(order), species = as.factor(species))

species_scores2 <- as.data.frame(nmds2[["species"]])

rownames(species_scores2) <- c("Optimal Temperature", "Length", "Lifespan", "Population Growth Rate")

species_scores2 <- species_scores2 %>% 
  rownames_to_column(var = "var")

centroid2 <- filled_scores2 %>%
  group_by(class2) %>%
  summarize(NMDS1 = mean(NMDS1), NMDS2 = mean(NMDS2))

nmds_3 <- ggplot(filled_scores2, aes(x = NMDS1, y = NMDS2)) +
  geom_vline(xintercept = 0, linetype = "dashed", color = "grey") +
  geom_hline(yintercept = 0, linetype = "dashed", color = "grey") +
  stat_ellipse(geom = "polygon", aes(color = class2, fill = class2), linewidth = 0.25, alpha = 0.2, show.legend = FALSE) +
  geom_point(aes(color = class2), alpha = 0.3) +
  theme_bw(base_size = 22) +
  geom_point(data = centroid2, aes(color = class2), size = 4, alpha = 0.9) +
  labs(x = "nMDS1", y = "nMDS2", color = "Movement Group (home range/PLD)") +
  scale_color_viridis_d() +
  scale_fill_viridis_d() +
  geom_segment(data = species_scores2, aes(x = 0, xend = MDS1, y = 0, yend = MDS2), arrow = arrow(length = unit(0.25, "cm")), linewidth = 0.5, colour = "grey20") +
  ggrepel::geom_label_repel(data = species_scores2, aes(x = MDS1, y = MDS2, label = var), size = 5, fill = "white", cex = 3, direction = "both", segment.size = .25) +
  scale_y_continuous(position = "right") +
  theme(legend.position = "left", panel.grid.major = element_blank(), panel.grid.minor = element_blank())

p2 = nmds_3

plot = p1 + p2 + plot_annotation(tag_levels = "A") + plot_layout(guides = "collect")

ggsave(plot, file = paste0("figS6.pdf"), path = here::here("03_figs"), height = 8, width = 18)


dst2 <- dist(com2, method = "euclidean")

bd2 <- betadisper(dst2, com_data2$class2)

anova(bd2)

permutest(bd2, pairwise = TRUE, permutations = 9999)

adonis2(com_data2$class3 ~ m_com2, permutations = 9999)

pairwise.adonis(m_com2, com_data2$class3, perm = 9999)
```

```{r}
filled_data = filled_data_save %>% 
  filter(!is.na(observed_homerange)) %>% 
  filter(!is.na(observed_pld))
```

```{r}
lm1 <- lm(log(home_range) ~ length * r_fin * longevity_wild * temp_preferred, data = filled_data_save %>% filter(!is.na(observed_homerange)))
lm2 <- lm(pld ~ length * r_fin * longevity_wild * temp_preferred, data = filled_data_save %>% filter(!is.na(observed_pld)))

summary(lm1)
AIC(lm1)
Anova(lm1, type = "III")

summary(lm2)
AIC(lm2)
Anova(lm2, type = "III")

tidy_lm1 <- tidy(lm1)
tidy_lm2 <- tidy(lm2)

write_csv(tidy_lm1, here::here("01_data", "02_processed_data", "hr_lm_empirical.csv"))
write_csv(tidy_lm2, here::here("01_data", "02_processed_data", "pld_lm_empirical.csv"))
```


```{r}
# optimal temperature -----------------------------------------------

temp_data <- filled_data %>%
  filter(!is.na(temp_preferred)) %>%
  filter(temp_preferred > 0)

temp_data %>% 
  group_by(class2) %>% 
  summarize(median(temp_preferred))

# look at outliers
outliers_temp <- boxplot.stats(temp_data$temp_preferred)$out
boxplot(temp_data$temp_preferred, main = "Preferred Temp", boxwex = 0.1)
mtext(paste("Outliers: ", paste(outliers_temp, collapse = ", ")), cex = 0.4)

# build the model
model_temp <- aov(temp_preferred ~ class2, data = temp_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_temp))
# not normal

# model with log transformed variable
model_temp_log <- aov(log(temp_preferred) ~ class2, data = temp_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_temp_log))
# more abnormal

# test for homogeneity of variance
plot(model_temp, 1)
# looks fine

# run Kruskal-Wallis test (non-parametric version of ANOVA)
kruskal.test(temp_preferred ~ class2, data = temp_data)
# chi-squared = 77.77, df = 5, p-value = 2.455e-15

# get effect size
temp_data %>% kruskal_effsize(temp_preferred ~ class2, ci = TRUE)
# eta2 = 0.124, MODERATE effect

# Kruskal-Wallis is significant, run Dunn's test
dunnTest(temp_preferred ~ class2, data = temp_data, method = "bh")

# length -------------------------------------------------------

# check for outliers - there are a lot?
outliers_length <- boxplot.stats(filled_data$length)$out
boxplot(filled_data$length, main = "Length", boxwex = 0.1)
mtext(paste("Outliers: ", paste(outliers_length, collapse = ", ")), cex = 0.4)

filled_data %>% 
  group_by(class2) %>% 
  summarize(mean = mean(length))

# build the model
model_length <- aov(length ~ class2, data = filled_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_length))
# not normal

# model with log transformed length
model_length_log <- aov(log(length) ~ class2, data = filled_data %>% filter(length > 0))

# create a QQ plot of residuals
ggqqplot(residuals(model_length_log))
# looks normal

# run Shapiro-Wilk test of normality
shapiro_test(residuals(model_length_log))
# p-value = 0.0669, normal

# test for homogeneity of variance
plot(model_length_log, 1)
# looks good

summary(model_length_log)
# p-value < 2e-16 ***

# effect size
effectsize::cohens_f(model_length_log)

effectsize::eta_squared(model_length_log)
interpret_eta_squared(0.53, rules = "cohen1992")
# eta2 = 0.18, LARGE effect

# post hoc Tukey test
TukeyHSD(model_length_log)

# lifespan -----------------------------------------------

longevity_data <- filled_data %>%
  filter(!is.na(longevity_wild))

longevity_data %>% 
  group_by(class2) %>% 
  summarise(median(longevity_wild))

# look at outliers
outliers_long <- boxplot.stats(longevity_data$longevity_wild)$out
boxplot(longevity_data$longevity_wild, main = "Longevity in the Wild", boxwex = 0.1)
mtext(paste("Outliers: ", paste(outliers_long, collapse = ", ")), cex = 0.4)

# build the model
model_long <- aov(longevity_wild ~ class2, data = longevity_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_long))
# not normal

# model with log transformed variable
model_long_log <- aov(log(longevity_wild) ~ class2, data = longevity_data %>% filter(longevity_wild > 0))

# create a QQ plot of residuals
ggqqplot(residuals(model_long_log))
# almost normal?

# run Shapiro-Wilk test of normality
shapiro_test(residuals(model_long_log))
# p-value = 0.00000807, not normal

# test for homogeneity of variance
plot(model_long, 1)
# looks fine?

# run Kruskal-Wallis test (non-parametric version of ANOVA)
kruskal.test(longevity_wild ~ class2, data = longevity_data)
# chi-squared = 25.928, df = 5, p-value = 9.216e-05

# get effect size
longevity_data %>% kruskal_effsize(longevity_wild ~ class2, ci = TRUE)
# eta2 = 0.0384, SMALL effect

# Kruskal-Wallis is significant, run Dunn's test
dunnTest(longevity_wild ~ class2, data = longevity_data, method = "bh")

# r_fin --------------------------------------------------------

# look at outliers
outliers_r_fin <- boxplot.stats(filled_data$r_fin)$out
boxplot(filled_data$r_fin, main = "Growth Rate", boxwex = 0.1)
mtext(paste("Outliers: ", paste(outliers_r_fin, collapse = ", ")), cex = 0.4)

filled_data %>% 
  group_by(class2) %>% 
  summarize(median(r_fin))

# build the model
model_r_fin <- aov(r_fin ~ class2, data = filled_data)

# create a QQ plot of residuals
ggqqplot(residuals(model_r_fin))
# not normal

# model with log transformed variable
model_r_fin_log <- aov(log(r_fin) ~ class2, data = filled_data %>% filter(r_fin > 0))

# create a QQ plot of residuals
ggqqplot(residuals(model_r_fin_log))
# normal?

# run Shapiro-Wilk test of normality
shapiro_test(residuals(model_r_fin_log))
# p-value = 0.000508, not normal

# test for homogeneity of variance
plot(model_r_fin, 1)
# looks fine

# run Kruskal-Wallis test (non-parametric version of ANOVA)
kruskal.test(r_fin ~ class2, data = filled_data)
# chi-squared = 42.245, df = 5, p-value = 5.255e-08

# get effect size
filled_data %>% kruskal_effsize(r_fin ~ class2, ci = TRUE)
# eta2 = 0.0621, MODERATE effect

# Kruskal-Wallis is significant, run Dunn's test
dunnTest(r_fin ~ class2, data = filled_data, method = "bh")

```

### Summary plot

```{r}

cohort <- c("1: HR-mid/PLD-low vs 2: HR-low/PLD-mid", "1: HR-mid/PLD-low vs 3: HR-mid/PLD-high", "1: HR-mid/PLD-low vs 4: HR-mid/PLD-very high", "1: HR-mid/PLD-low vs 5: HR-mid/PLD-mid", "1: HR-mid/PLD-low vs 6: HR-high/PLD-mid",
            "2: HR-low/PLD-mid vs 3: HR-mid/PLD-high", "2: HR-low/PLD-mid vs 4: HR-mid/PLD-very high", "2: HR-low/PLD-mid vs 5: HR-mid/PLD-mid", "2: HR-low/PLD-mid vs 6: HR-high/PLD-mid",
            "3: HR-mid/PLD-high vs 4: HR-mid/PLD-very high", "3: HR-mid/PLD-high vs 5: HR-mid/PLD-mid", "3: HR-mid/PLD-high vs 6: HR-high/PLD-mid",
            "4: HR-mid/PLD-very high vs 5: HR-mid/PLD-mid", "4: HR-mid/PLD-very high vs 6: HR-high/PLD-mid",
            "5: HR-mid/PLD-mid vs 6: HR-high/PLD-mid")
PERMDISP <- c(0.025, 0.028, 0.714, 0.618, 0.913, 
              0.0648, 0.696, 0.555, 0.471, 
              0.395, 0.218, 0.313, 
              0.985, 0.8464, 
              0.822)
PERMANOVA <- c(1.00, 0.003, 0.0015, 0.0855, 0.0135, 
              0.003, 0.0480, 0.3285, 0.0210,
              1.00, 0.054, 0.0885, 
              0.0195, 0.138, 
              1.00)
Optimal_Temp <- c(0.156, 0.073, 0.0062, 0.4705, 0.0323, 
              0.1772, 0.0100, 0.1494, 0.0063, 
              0.2105, 0.05855, 0.0043, 
              0.00339, 0.00028, 
              0.1562)
temp_arrows <- c("increase", "increase", "increase", "decrease", "decrease", 
              "increase", "increase", "decrease", "decrease", 
              "increase", "decrease", "decrease", 
              "decrease", "decrease", 
              "decrease") # decrease or increase
Length <- c(0.614, 0.998, 0.209, 0.000006, 0.0000003, 
              0.8909, 0.482, 0.0000595, 0.0000019, 
              0.29, 0.0013, 0.0000212, 
              0.759, 0.0894, 
              0.522)
length_arrows <- c("decrease", "increase", "decrease", "decrease", "decrease", 
              "increase", "decrease", "decrease", "decrease", 
              "decrease", "decrease", "decrease", 
              "decrease", "decrease", 
              "decrease")
Lifespan <- c(0.75, 0.322, 0.5404, 0.615, 0.666, 
              0.1756, 0.536, 0.655, 0.604, 
              0.634, 0.523, 0.136, 
              0.5999, 0.5100, 
              0.478)
life_arrows <- c("decrease", "decrease", "decrease", "decrease", "increase", 
              "decrease", "decrease", "decrease", "increase", 
              "increase", "increase", "increase", 
              "increase", "increase", 
              "increase")
Growth_Rate <- c(0.342, 0.0119, 0.359, 0.319, 0.402, 
              0.0475, 0.513, 0.602, 0.654, 
              0.478, 0.352, 0.393, 
              0.822, 0.922, 
              0.868)
growth_arrows <- c("increase", "increase", "increase", "increase", "increase", 
              "increase", "increase", "increase", "increase", 
             "decrease" ,"decrease" ,"decrease" , 
              "decrease","decrease" , 
              "increase")


# Create data frame
data <- data.frame(cohort, PERMDISP, PERMANOVA, Optimal_Temp, Length, Lifespan, Growth_Rate)

# Pivot to long data format
data_long <- data %>% 
  mutate(across(c("Optimal_Temp", "Length", "Lifespan", "Growth_Rate"), round, 4)) %>% 
  pivot_longer(-cohort) %>% 
  set_names(c("cohort", "variable", "p_value")) %>% 
  mutate(significant = case_when(
    p_value >= 0.05 ~ "no",
    p_value < 0.05 ~ "yes"
  ))

# Add arrow information
data1 <- data.frame(cohort, temp_arrows, length_arrows, life_arrows, growth_arrows) %>% 
  rename(Optimal_Temp = temp_arrows, Length = length_arrows, Lifespan = life_arrows, Growth_Rate = growth_arrows) %>% 
  pivot_longer(-cohort) %>% 
  set_names(c("cohort", "variable", "direction")) %>% 
  full_join(data_long, by = c("cohort","variable")) %>% 
  mutate(direction = fct_relevel(direction, c("decrease", "no change", "increase")))

# Reorder and rename levels
data1$variable <- factor(data1$variable, levels = c("PERMANOVA", "PERMDISP", "Optimal_Temp", "Length", "Lifespan", "Growth_Rate"))
data1$variable <- recode_factor(data1$variable, PERMANOVA = "PERMANOVA", PERMDISP = "PERMDISP", Optimal_Temp = "Optimal Temperature", Length = "Length", Lifespan = "Lifespan", Growth_Rate = "Growth Rate")



# Plot 
g <- ggplot(data1, aes(x = variable, y = cohort, fill = direction)) +
  geom_tile(color = "white", lwd = 1) +
  geom_point(data = data1 %>% filter(significant == "yes"), shape = 8, size = 3, color = "black", show.legend = FALSE) +
  theme_minimal(base_size = 12) +
  theme(axis.title = element_blank(),
        axis.text = element_text(face = "bold", color = "black"),
        axis.ticks = element_blank(),
        panel.grid.major = element_blank()) +
  labs(fill = "Direction of Effect") +
  scale_x_discrete(position = "top") +
  scale_y_discrete(limits = rev) +
  scale_fill_manual(values = c( "#7568a0", "#fde725"), na.value = "darkgray")


```

```{r}

cohort <- c("1: HR-mid/PLD-low vs 2: HR-low/PLD-mid", "1: HR-mid/PLD-low vs 3: HR-mid/PLD-high", "1: HR-mid/PLD-low vs 4: HR-mid/PLD-very high", "1: HR-mid/PLD-low vs 5: HR-mid/PLD-mid", "1: HR-mid/PLD-low vs 6: HR-high/PLD-mid",
            "2: HR-low/PLD-mid vs 3: HR-mid/PLD-high", "2: HR-low/PLD-mid vs 4: HR-mid/PLD-very high", "2: HR-low/PLD-mid vs 5: HR-mid/PLD-mid", "2: HR-low/PLD-mid vs 6: HR-high/PLD-mid",
            "3: HR-mid/PLD-high vs 4: HR-mid/PLD-very high", "3: HR-mid/PLD-high vs 5: HR-mid/PLD-mid", "3: HR-mid/PLD-high vs 6: HR-high/PLD-mid",
            "4: HR-mid/PLD-very high vs 5: HR-mid/PLD-mid", "4: HR-mid/PLD-very high vs 6: HR-high/PLD-mid",
            "5: HR-mid/PLD-mid vs 6: HR-high/PLD-mid")
PERMDISP <- c(0.596, 0.5006, NA, 0.207, NA, 
              0.382, NA, 0.449, NA, 
              NA, 0.105, NA, 
              NA, NA, 
              NA)
PERMANOVA <- c(1, 1, 1, 1, 1, 
              1, 1, 1, 0.97, 
              1, 1, 1, 
              1, NA, 
              1)
Optimal_Temp <- c(0.415, 0.359, 0.297, 0.818, 0.388, 
              0.322, 0.408, 0.573, 0.309, 
              0.360, 0.451, 0.873, 
              0.295, 0.233, 
              0.406)
temp_arrows <- c("increase", "decrease", "increase", "increase", "decrease", 
              "decrease","increase" , "decrease", "decrease", 
              "increase", "increase", "decrease", 
              "decrease", "decrease", 
              "decrease") # decrease or increase
Length <- c(0.094, 0.828, 0.691, 0.964, 0.493, 
              0.927, 0.502, 0.748, 0.318, 
              0.327, 0.525, 0.207, 
              0.928, 0.999, 
              0.799)
length_arrows <- c("increase", "increase", "decrease", "decrease", "decrease", 
              "increase", "decrease", "decrease", "decrease", 
              "decrease", "decrease", "decrease", 
              "increase", "decrease", 
              "decrease")
Lifespan <- c(1, 1, 1, 1, 1, 
              1, 1, 0969, 1, 
              1,1 , 1, 
              1, 1, 
             1 )
life_arrows <- c("decrease", "decrease", "decrease","decrease" , "increase", 
              "decrease", "increase", "decrease", "increase", 
              "increase", "increase", "increase", 
              "decrease", "increase", 
              "increase")
Growth_Rate <- c(1, 1,1 , 1,0.969 , 
              1, 1, 1, 1, 
              1,1 , 1, 
              0.999,0.996 , 
             1 )
growth_arrows <- c("decrease", "decrease", "increase", "increase", "increase", 
              "decrease", "increase", "increase", "increase", 
              "increase", "increase", "increase", 
              "increase", "decrease", 
              "decrease")

# Create data frame
data <- data.frame(cohort, PERMDISP, PERMANOVA, Optimal_Temp, Length, Lifespan, Growth_Rate)

# Pivot to long data format
data_long <- data %>% 
  mutate(across(c("Optimal_Temp", "Length", "Lifespan", "Growth_Rate"), round, 4)) %>% 
  pivot_longer(-cohort) %>% 
  set_names(c("cohort", "variable", "p_value")) %>% 
  mutate(significant = case_when(
    p_value >= 0.05 ~ "no",
    p_value < 0.05 ~ "yes"
  ))

# Add arrow information
data1 <- data.frame(cohort, temp_arrows, length_arrows, life_arrows, growth_arrows) %>% 
  rename(Optimal_Temp = temp_arrows, Length = length_arrows, Lifespan = life_arrows, Growth_Rate = growth_arrows) %>% 
  pivot_longer(-cohort) %>% 
  set_names(c("cohort", "variable", "direction")) %>% 
  full_join(data_long, by = c("cohort","variable")) %>% 
  mutate(direction = fct_relevel(direction, c("decrease", "no change", "increase")))

# Reorder and rename levels
data1$variable <- factor(data1$variable, levels = c("PERMANOVA", "PERMDISP", "Optimal_Temp", "Length", "Lifespan", "Growth_Rate"))
data1$variable <- recode_factor(data1$variable, PERMANOVA = "PERMANOVA", PERMDISP = "PERMDISP", Optimal_Temp = "Optimal Temperature", Length = "Length", Lifespan = "Lifespan", Growth_Rate = "Growth Rate")



# Plot 
g2 <- ggplot(data1, aes(x = variable, y = cohort, fill = direction)) +
  geom_tile(color = "white", lwd = 1) +
  geom_point(data = data1 %>% filter(significant == "yes"), shape = 8, size = 3, color = "black", show.legend = FALSE) +
  theme_minimal(base_size = 12) +
  theme(axis.title = element_blank(),
        axis.text = element_text(face = "bold", color = "black"),
        axis.ticks = element_blank(),
        panel.grid.major = element_blank()) +
  labs(fill = "Direction of Effect") +
  scale_x_discrete(position = "top") +
  scale_y_discrete(limits = rev)  +
  scale_fill_manual(values = c( "#7568a0", "#fde725"), na.value = "darkgray")

plot = g + g2 + plot_annotation(tag_levels = "A") + plot_layout(guides = "collect")

ggsave(plot, file = paste0("figS7.pdf"), path = here::here("03_figs"), height = 12, width = 25)
```
